fix: 64-bit row indexing in w8a16 output store (int32 overflow at large prefill M)
Companion to the swiglu int64 fix (5d3c9ad). The w8a16 GEMM output store
offs_ym*stride_ym (= M*N, up to ~3e9 for M~1e5, N=28672) exceeds int32
range and wraps to a negative address -> illegal memory access on large
prefill. This is the version that was profiled for the EXP11 result
(4924 agg_tps); it was inadvertently left out of the EXP11 commit d298ba9.

+++ kernels/w8a16_gemm_kernel.py
@@ -94,7 +94,11 @@ def _w8a16_gemm(
     acc = acc * scale[None, :]
 
     offs_ym = pid_m * BLOCK_M + tl.arange(0, BLOCK_M)
-    y_ptrs = y_ptr + offs_ym[:, None] * stride_ym + offs_yn[None, :] * stride_yn
+    # 64-bit row offset: for large prefill M the product offs_ym*stride_ym (= M*N)
+    # can exceed int32 range (e.g. M~1e5, N=28672 -> ~3e9 > 2^31), which would wrap
+    # to a negative address and trigger an illegal memory access. Promote to int64.
+    y_ptrs = (y_ptr + offs_ym[:, None].to(tl.int64) * stride_ym
+              + offs_yn[None, :].to(tl.int64) * stride_yn)
     y_mask = (offs_ym[:, None] < M) & (offs_yn[None, :] < N)
     tl.store(y_ptrs, acc.to(y_ptr.dtype.element_ty), mask=y_mask)
